float head mask

+++ generate_dropout.py
@@ -114,7 +114,7 @@ def _main(cfg: DictConfig, output_file):
         num_shards=cfg.checkpoint.checkpoint_shard_count,
     )
 
-    head_mask = convert_gate_to_mask(models[0].get_w(), saved_cfg.pruning.num_of_heads)
+    head_mask = convert_gate_to_mask(models[0].get_w(), saved_cfg.pruning.num_of_heads).float()
     head_mask = utils.move_to_cuda(head_mask) if use_cuda else head_mask
     models[0].apply_masks(head_mask)
